Make wxQtEnsureSignalsBlocked available to all wxQt classes
and use it instead of using a pair of calls to blockSignals()
