+++ src/qt/calctrl.cpp
@@ -137,9 +137,8 @@ bool wxCalendarCtrl::SetDate(const wxDateTime& date)
             wxQtConvertDate( date ) < m_qtCalendar->minimumDate() )
         return false;
 
-    m_qtCalendar->blockSignals(true);
+    wxQtEnsureSignalsBlocked blocker(m_qtCalendar);
     m_qtCalendar->setSelectedDate(wxQtConvertDate(date));
-    m_qtCalendar->blockSignals(false);
 
     return true;
 }
@@ -158,10 +157,9 @@ bool wxCalendarCtrl::SetDateRange(const wxDateTime& lowerdate,
     if ( !m_qtCalendar )
         return false;
 
-    m_qtCalendar->blockSignals(true);
+    wxQtEnsureSignalsBlocked blocker(m_qtCalendar);
     m_qtCalendar->setMinimumDate(wxQtConvertDate(lowerdate));
     m_qtCalendar->setMaximumDate(wxQtConvertDate(upperdate));
-    m_qtCalendar->blockSignals(false);
 
     return true;
 }
