+++ src/qt/checklst.cpp
@@ -9,6 +9,7 @@
 #include "wx/wxprec.h"
 
 #include "wx/checklst.h"
+#include "wx/qt/private/winevent.h"
 
 #include <QtWidgets/QListWidgetItem>
 
@@ -81,12 +82,10 @@ bool wxCheckListBox::IsChecked(unsigned int n) const
 void wxCheckListBox::Check(unsigned int n, bool check )
 {
     // Prevent the emission of wxEVT_CHECKLISTBOX event by temporarily block all
-    // signals on m_qtListWidget object. QSignalBlocker can be used instead when
-    // Qt 5.3 becomes the minimum supprted version.
-    const bool wasBlocked = m_qtListWidget->blockSignals(true);
+    // signals on m_qtListWidget object.
+    wxQtEnsureSignalsBlocked blocker(m_qtListWidget);
     QListWidgetItem* item = m_qtListWidget->item(n);
     wxCHECK_RET(item != nullptr, wxT("wrong listbox index") );
     item->setCheckState(check ? Qt::Checked : Qt::Unchecked);
-    m_qtListWidget->blockSignals(wasBlocked);
 }
