+++ src/qt/spinctrl.cpp
@@ -67,17 +67,15 @@ wxString wxSpinCtrlQt< T, Widget >::GetTextValue() const
 template< typename T, typename Widget >
 void wxSpinCtrlQt< T, Widget >::SetValue( T val )
 {
-    m_qtSpinBox->blockSignals(true);
+    wxQtEnsureSignalsBlocked blocker(m_qtSpinBox);
     m_qtSpinBox->setValue( val );
-    m_qtSpinBox->blockSignals(false);
 }
 
 template< typename T, typename Widget >
 void wxSpinCtrlQt< T, Widget >::SetRange( T min, T max )
 {
-    m_qtSpinBox->blockSignals(true);
+    wxQtEnsureSignalsBlocked blocker(m_qtSpinBox);
     m_qtSpinBox->setRange( min, max );
-    m_qtSpinBox->blockSignals(false);
 }
 
 template< typename T, typename Widget >
